+++ CHANGELOG.md
@@ -1,5 +1,8 @@
+###v. 0.4.5 (23 Oct 2017)
+While processing nested objects, we should handle dates accordingly (https://github.com/yury-dymov/json-api-normalizer/issues/23)
+
 ###v. 0.4.4 (23 Oct 2017)
-While we process nested objects, we shouldn't change array attributes to object (https://github.com/yury-dymov/json-api-normalizer/issues/22)
+While processing nested objects, we shouldn't change array attributes to object (https://github.com/yury-dymov/json-api-normalizer/issues/22)
 
 ###v. 0.4.3 (20 Oct 2017)
 Nested attribute keys are also camelized now (https://github.com/yury-dymov/json-api-normalizer/issues/21)